remove allowed-network confg
This was due to misunderstanding of requirements. Now the flow to get our connection info is:
- If a hard-coded contact is present use the Endpoint Echo service to querry.
- If no hard-coded contacts are present (because we might be the genesis peer), query the underlying socket for address
- If the address obatained is of unspecified category then ignore the result and error out as no one can connect to an unspecified address so it's not useful for the purposes of giving out connection info to others

So basically if you are a genesis peer, you can omit specifying the port number in the config file in which case a random OS allocated port will be used, but you will have to specify an IP address. Then the connection info will be given/printed with the specified IP address (as the socket will have bound to that), the OS allocated random port (or the supplied port if any via the config) which the socket has bound to and the random self-singed certificate (unless that was supplied in the config too in which case the supplied certificate will be used).