Reduce test data size
We used to send 400 MB of data. This data would be serialized before put
on the wire. Serialization process consumes a lot of CPU which blocks
event loop thread.  If blocked for too long, it result in remote stream
timeouts. Reducing test data to 10 MB makes serialization faster and
avoids mentioned problems and makes the tests pass.