use QConn wrapper
This makes sure we have RAII for destroying quinn::Connection instead of tracking it in many places in code making it error prone. The upstream behaviour is that it will only call close if there are no references to the connection such as existing unresolved IncomingStreams etc. which is very error prone to track in our case and we want the connection gone on forgetting our handle which is the Connection object with Context. Thus RAII makes code far simpler and, more importantly, robust in our case.