fix/wire_msg: fix incorrect serialisation logic in WireMsg
While deserialising messages, we wrongly assumed that all messages >1 KB in size
are not serialised and are plain user messages. However, bincode::serialise(wire_msg)
also could have produced messages over 1 KB, which resulted in deserialisation failures
in the user code.

To fix this, we added a one-byte flag that will clearly identify pre-serialised user messages
and messages that are sent by quic-p2p itself. With that, the serialisation logic becomes even
more efficient now because we don't need to serialise _all_ user messages which are likely to
be serialised already.