fix/event-loop: force the use of basic tokio scheduler
If quic-p2p was included as a dependency in a repo with tokio dep listing
a rt-threaded feature, it defaulted to use the threaded scheduler, resulting
in a failure of finding the quic-p2p context (which is a thread local variable).
This commit provides a fix for this problem by forcing the use of the basic,
single-threaded tokio scheduler.