fix(bootstrap): fix stalled connections w/multiple bootstrap contacts
When multiple bootstrap contacts are used the first successful
connection is considered and the rest is disregarded. This results in
unfinished connection attempts that are left lingering in the connection
de-duplicator and connection pool. This fix introduces
Endpoint::new_connection that creates a fresh connection
disregarding the de-duplication of connections that is used by the
bootstrap function which then only saves the successful connection in the pool