feat: Make total backoff duration configurable
This is how the `backoff` crate supports limiting how long we back off
for. The default is 15 mins (!) which is crazy long for our purposes.

The configuration is exposed via the `Config` struct under
`retry_duration_msec`. The default is 30s, which is completely arbitrary
but reasonable-sounding. Retry logic was refactored into a method to
make it easier to use the configuration.