test: Synchronise quinn test close on sending completion
We are seeing occasional test failures due to a race condition between
the sending and receiving sides of the clients (`sender`s) in this test.
Specifically, it's possible for the receiving side to finish before the
sending side finishes. When this happens, the sending side crashes when
the receiving side closes the connection.

The simplest way to prevent this is to synchronise the call to
`sender.close` on the sending side actually finishing. This has been
implemented using `tokio::sync::Notify`.