refactor!: Remove random port fallback
This is a bit 'too clever' for this library, and it's unlikely that this
fallback was really being used (it would usually be left unspecified, so
always be random).

BREAKING CHANGE: This is a breaking behavioural change. It is now up to
the caller if they want to retry with a different port.