refactor!: Refactor `WireMsg::read_from_stream` to avoid logic error
Previously we used `TryFrom` which we know will succeed thanks to our
assertions, however we don't want to allow for the possibility of
unexpected panics by using `unwrap` here.

Since we have asserted the pointer size though, we can reasonably
perform a lossy conversion with `as`, knowing that in fact nothing will
be lost.

BREAKING CHANGE: The `Error::UnexpectedError` variant has been removed.