refactor!: Make `ConnId::generate` infallible
This trait method originally allowed for an arbitrary `Box<dyn Error>`,
likely to account for the possible serialization failure when using
`bincode::serialize` in the impl for `XorName`. Whilst this allows for
more exotic connection ID strategies, having to account for this error
in all APIs that touch connection IDs is a bit of a pain.

By rewriting the `XorName` impl to 'manually' extract the IP and port
bytes, we can avoid any possibility of error and make the method
infallible.

BREAKING CHANGE: The `ConnId::generate` method now returns `Self`, and
doesn't allow for errors. The `Error::ConnectionIdGeneration` variant
has been removed.