feat: Add `Endpoint::new_client` constructor
This creates an `Endpoint` that can work as a client only (all the
machinery for incoming transactions are dropped).

Another difference from `Endpoint::new` is that this does none of the
reachability checking, since this should not be necessary for client-
only endpoints (though, the API for bidirectional streams is currently
unusable). This massively reduces the possible errors, which have their
own `ClientError` enum.