refactor: Remove `connection_deduplicator::Error`
This introduces a new (crate private) enum `ConnectFailure` which covers
the three possible reasons a connect attempt might fail:

- The connect parameters were invalid (`quinn::ConnectError`).
- An I/O or protocol error occurred during connect
  (`quinn::ConnectionError`), which we expose as two variants in our
  API:
  - `Error::ConnectionClosed` indicating the connection was closed
    gracefully by ourselves, the application, or the protocol.
  - `Error::ConnectionError` for other possibilities (resets, timeouts,
    and transport errors).

Ultimately, this allows us to ditch `connection_deduplicator::Error` in
favour of `ConnectFailure` which can be used for other things in future.

The separation of `quinn::ConnectionError` to `Close` and
`ConnectionError` has also be factored into a private
`ConnectionCloseOrError` enum.