refactor!: Merge 'connect' and 'close' errors with `ConnectionError`
This introduces new variants to the `ConnectionError` enum to cover
failures that may be encountered before establishing a connection
(`quinn::ConnectError`) as well as closures (`Close`).

This is useful for a couple of reasons:

- Many of our APIs will transparently try to create connections before
  using them. Even when reducing the possible variants from these APIs,
  we would still need to account for these three possibilities
  (pre-connect error, connection error, and connection close).
- It removes `quinn::ConnectError` from our public API, and better
  represents errors that should be impossible as bugs.
- It replaces the crate-private `ConnectFailure` and simplifies some of
  the error conversion logic.

BREAKING CHANGE: The `Error::Connect` and `Error::ConnectionClosed`
variants have been removed. Several additional variants have been added
to `ConnectionError`.