refactor!: Remove `Error::UnknownStream` variant
This variant existed to handle errors returned by quinn when trying to
set or get the stream's priority. The errors can occur because the
stream objects are essentially indices into a map of underlying streams.
Calling methods on the stream translates to a lookup into this map.
Undelrying streams may close for various reasons independent of our
handles for them (e.g. if the stream or connection closes).

However, handling this situation as errors is not very useful. If we
fail to set the priority of a stream because it doesn't exist, the
effect of the priority would be unobservable. Any attempt to observe the
effect (e.g. reading/writing to the stream) would lead to an error at
that point, which is a far more convenient place to handle them.

By *ignoring* errors when setting stream priority (and dropping the
getter for priority, since there's no downstream use-case), we can
remove the `UnknownStream` variant and reduce the range of possible
errors from all functions that establish connections (which is many!).

BREAKING CHANGE:

- The `Error::UnknownStream` variant has been removed.
- `SendStream::set_priority` is now infallible and returns `()`.
- `SendStream::priority` has been removed.