refactor: Centralise connection pool interactions
`Endpoint` had a few different methods for connecting, some of which
would use the connection pool + deduplicator and some would not. It's
not really clear when we wouldn't want to retry connection attempts, so
this commit introduces a (private) `get_or_connect_to` function which
always uses the pool and always retries. All functions that need a
connection now use this.

This has the nice side-effect of removing (most of) the need for
`Error::MissingConnection`, since `get_or_connect_to` returns the
established connection.