refactor!: Treat unexpected messages as serialization errors
The semantics of this are a bit questionable, but it allows us to remove
the `RecvNextError` and `UnexpectedMessageType` types. Essentially, we
treat the problem of receiving an unexpected message as if we've tried
to deserialize the expected message, and found something else. This
doesn't quite match the logic (we in fact *do* deserialize a message
successfully), but is suitable for the API.

BREAKING CHANGE: The `RecvNextError` and `UnexpectedMessageType` types
have been removed. `RecvError` is used where `RecvNextError` was used
previously.