refactor!: Return `SendError` in `Endpoint::try_send_message`
Rather than returning an `Error::MissingConnection` if the connection is
not in the pool, this now returns `Err(None)`. If the connection *is* in
the pool, but there's an error sending the message, this will return
`Err(Some(error))`. It remains to be seen how usable this will be for
callers, but it avoid introducing another type for this single case and
is a reasonable model.

BREAKING CHANGE: `Endpoint::try_send_message` now returns `Result<(),
Option<SendError>>` rather than `Result<(), Error>`. The
`Error::MissingConnection` variant has been removed.