fix: Don't report an error on benign end of stream
The implementation of reading multiple messages from a stream involved
looping until an error was encountered, and this error would then be
logged. We typically send just a single message per stream, so this
implementation meant we would log a serializatione error after reading
every message.

This commit updates `WireMsg::read_from_stream` to return an `Option`,
and `None` is now used to break the loop. No public APIs have changed –
APIs that must return a message will still return a serialization error
if the stream ends (though the error message has changed, but this is
judged to be outside the stable API).