refactor!: Make `Config::port_forward` depend on `igd` feature
The `Config` struct will no longer have a `port_forward` field if the
`igd` feature is not enabled. This removes a case where a user could set
the config, but we would ignore it.

BREAKING CHANGE: The `port_forward` field on `Config` is only present
with the `igd` feature enabled (default).