refactor!: yield `Connection`s from `Incoming*` streams
This is a necessary API change on the path to allowing the
`ConnectionPool` to be removed – it's would no longer be sufficient to
yield the `SocketAddr`, since there would be no way of obtaining the
associated `Connection`.

Since the `ConnectionPool` yet remains, this is again an easy migration
for callers, who can simply use `Connection::remote_address` and work
with the pre-existing APIs.

BREAKING CHANGE: `IncomingConnections::next` now returns
`Option<Connection<I>>`. `IncomingMessages::next` now returns
`Option<(Connection<I>, Bytes)>`.