test: return errors in tests, don't just construct them
This is a bug that's been sitting for a while, and was even maintained
in the conversion from `anyhow` -> `eyre`.

Thankfully it was masking a single test failure, related to the
behaviour of `disconnect_from` which was deliberately changed to close
all connections to the peer back in 2bab0548, so this was just a case of
updating the test.