feat: add a counter for opened connections to `Endpoint`
This is behind an `unstable-opened-connection-count` feature flag, so
that it can be removed again in future without a breaking change. The
intention is to use it to get a measure of the current number of
connections that we open in downstream tests, in order to see the effect
of removing the `ConnectionPool` and track down extraneous connections.