refactor: implement a non-pooled `Connection`
This commit adds a `connection` module, which implements a connection
API with no dependencies on the connection pool. Most of the API has
moved from `connection_handle`, but from now `ConnectionHandle` is
still exported as `Connection`.

When used on its own, the new `Connection` API follows the typical
drop-based lifecycle. Once both sides of the API (`Connection` and
`ConnectionIncoming`) are dropped, the connection will close
automatically. However, for compability in the short term the `Endpoint`
APIs will still put these `Connection`s in the connection pool, and poll
for incoming messages in a background task, so for now connections will
still remain open until error or explicit close.