refactor!: remove connection pooling
This is a significant breaking change to the library's API and
behaviour. In particular:

- All the message-oriented APIs have been removed from `Endpoint`,
  including `IncomingMessages`. You can use `Connection` and
  `ConnectionIncoming` from `Endpoint::connect_to` or
  `IncomingConnections`.
- `DisconnectionEvents` has been removed, and is no longer returned from
  `Endpoint::new*`. Remote disconnections can be detected when
  performing operations on a `Connection` returns a `ConnectionError`.
  Dropping both halves of a connection (`Connection` and
  `ConnectionIncoming`) will immediately disconnect locally.
- Connections are not held open automatically. If the
  `IncomingConnections` struct is dropped when creating an `Endpoint`,
  any connections handled by the endpoint will be dropped immediately
  (in future this should probably drop the underlying `quinn::Incoming`
  as well, meaning connections will be rejected outright).

The upshot of this is change is that it's now necessary for callers to
apply their own resource management for connection handles. A naive
migration would likely lead to many more connections being made, which
could be detrimental to performance. However, this simplifies `qp2p` and
removes a lot of edge-cases, and is expected to be removing a lot of
undiscovered bugs. Callers can now use a connection management strategy
that makes the most sense for them, which is likely to be much simpler
than the general-purpose approach that had to be taken here.

BREAKING CHANGE: Many `Endpoint` APIs have been removed or changed.