refactor!: remove `DisconnectionEvents`
This removes the `DisconnectionEvents` API, and all the plumbing that
makes it work. This is part of the move to a fully connection-oriented
API. In that case, callers would always be working with a handle to a
connection, and so will see if disconnections occur.

BREAKING CHANGE: `DisconnectionEvents` has been removed, and is no
longer returned from `Endpoint::new` or `Endpoint::new_client`.