refactor!: remove `IncomingMessages`
This is part of the move to a fully connection-oriented API. The
per-connection `ConnectionIncoming` streams must be used instead.

Since `ConnectionIncoming` cannot be cloned (since the inner
`mpsc::Receiver` cannot be cloned), methods that use the connection pool
return an `Option<ConnectionIncoming>`. This also reveals one of the
inconsistencies with the connection pool, in that incoming connections
always come with a `ConnectionIncoming`, since they replace anything
already in the pool.

As with `Connection`, it's currently a connection-pool-specific wrapper
that's returned (`ConnectionIncomingHandle`). This is necessary to
ensure that lost connections are removed from the pool. These wrappers
will go away when the connection pool is removed, and the API should
remain unchanged.

BREAKING CHANGE: `IncomingMessages` has been removed, and is no longer
returned by `Endpoint::new` or `Endpoint::new_client`.