fix: more carefully handle 'benign' connection loss
This changes the behaviour of the `ConnectionIncoming` pseudo-stream to
signal end-of-stream when the connection closes for a 'benign' reason,
rather than returning an error. This makes it much easier to implement
message polling for callers, without them having to apply their own
filtering.

Since we're doing this detection in a couple of places, it made sense to
move the check into `ConnectionError` itself.