feat: add `Connection::id` to get a stable ID for a connection
This just proxies the underlying `quinn::Connection::stable_id`. This is
actually implemented in `quinn` as the address in memory of the
underlying shared connection reference, so although it is stable for the
lifetime of a connection it is not comparable across peers. It's still
useful for tracking connections within a single peer.