test: increase timeout for some tests
We want tests to fail, rather than hanging forever, in situation where
we expect some incoming connection/message. Sadly, setting this timeout
is a bit arbitrary, and depending on the performance of the system
running the tests they could fire spuriously. That seems to be what
we're seeing in CI, so we're bumping the timeout to compensate.

Note that this will make some tests take longer, since the timeout is
also used to check that something we don't expect to happen, doesn't
happen.