chore: bump edition to 2021
The only change that affected us was that the Cargo resolver has changed
to be version 2 by default. It turns out that this causes compiling
without dev-dependencies to fail, since we were depending on
`tiny-keccak` without specifying any hash function features. We don't
actually use `tiny-keccak` in the library any more, just in the tests,
so this has simple been removed from the main dependencies.