refactor!: improve retry APIs
This is an attempt to allow callers to maintain a 'retry context'
throughout their own code and calls to `qp2p`. For example, it may be
desired to retry an operation for 60s, regardless of where the failure
occurs. Without a change such as this, callers would have to wrap their
entire operation in a top level retry, and it becomes difficult to
reason about where retries may occur and for how long.

The APIs introduced here allow callers to supply an implementation of
`backoff::backoff::Backoff`, which will be used to drive retries. That
alone is insufficient to allow a 'retry context' to be maintained, but
the included `qp2p::backoff::BackoffRef` utility struct allows a mutable
reference to a `backoff::backoff::Backoff` to be passed.

BREAKING CHANGE: `Connection::send_with` now takes an `impl Backoff`
rather than an `Option<&RetryConfig>` for the 3rd argument.