feat: add 'structopt' feature flag
structopt was removed in edc5493294e6ede45ebbd0023be51757c355793d.

However it is very useful for cli programs to be able to
expose p2p options on the command-line and read in a
qp2p Config easily.

This commit adds back structopt support for Config, now
as an optional feature flag, 'structopt'.