refactor(api)!: refactoring and addition of new Error types to the public API
- In many cases the API was returning a `SerialisationError` type upon receiving messages
of unexpected type, thus these errors were misleading. This PR introduces specific error
types for those scenarios which shall make it clearer for the caller, also enabling proper
error handling for those cases.
- The `RecvStream::next` public API now returns `Result<Option<UsrMsgBytes>, RecvError>`
instead of `Result<UsrMsgBytes, RecvError>` which allows the caller to realise when the
stream was finished as opposed to any error that could occur when attempting to read from it.