fix(api): don't treat any type of errors as benign
BREAKING CHANGE: Currently when an error occurs upon trying to read from a stream,
we swallow some of the errors assuming them benign, as if the connection
was just nicely closed. It'll be better to report the error upwards so
the caller/user of this lib can be aware of it since it knows much better
the context where it's happening to make any decision.